refactor(ports): expand core/ports/tts.py with SynthesisResult, TtsUnavailableError
Add SynthesisResult dataclass and TtsUnavailableError inline into core/ports/tts.py.
Remove TYPE_CHECKING cross-reference to lyra.tts. Update lyra.tts re-export shim to
source SynthesisResult and TtsUnavailableError from core/ports/tts. Dead code
(TTSConfig, load_tts_config) not included in core/ports.
